SteamArtwork extract onDeleteFile, onCopyFile
