+++ assets_dev/loca/loca.csv
@@ -1,3 +1,7 @@
 keys;en
 homeCleanUpCacheDescription;There is unused art in your cache. You can free up {value}.
 homeCleanUpCacheButton;Clean up
+homeProgramsEmpty;No programs have been added to steam launcher.
+homeProgramArtworkDelete;Delete
+homeProgramArtworkSetBackgroundAsHero;Set as hero
+homeProgramArtworkSetHeroAsBackground;Set as background
